Add Robo bootstrap script generation and saving.
A user can record a sequence of actions using Test Recorder
and then save it as a Robo bootstrap script in JSON format
(such that it is readable and editable). The bootstrap script
can be used to guide the Robo's app exploration, e.g., to
reach a particular activity, to complete the login process, etc.

Test: Covered by test com.google.gct.testrecorder.ui.RecordingDialogTest
Change-Id: I37e8e3e6c7670408e651b6a7a854d648e5d353e1