Support permissions requested during a recording session.
Starting with version 1.0.0, ATSL introduced GrantPermissionRule that
allows to grant all the required permissions while running an Espresso
test such that it is not blocked by permission request dialogs on API
23+.

This CL makes Espresso Test Recorder detect permission requests during a
recording session and add the corresponding GrantPermissionRule to the
generated test code. Also, it would detect and update ATSL version to
make sure it is new enough to support GrantPermissionRule.

Fixes: b/37132717

Test: Covered by TestCodeMapperTest and TestCodeGeneratorTest.
Change-Id: Ibd6ff952d78ad3c7fd61114e9f18ec485807d946