Avoid reformatting the initially generated template file.
The initially generated template file has Java content, so when it is
generated for a Kotlin project, and the reformatter is triggered
before Espresso Test Recorder writes out the content of the generated
test class in Kotlin, the reformatter throws an IDE exception.

Also, explicitly reformat the generated test code after making sure that
import optimizer has finished (otherwise, import optimizer fails to
optimize the imports).

Fixes: 183648652
Test: Manual
Change-Id: Ib9b533e05d5edbcc885a4e57f2327f34ec543bf3