Replace Activating and Deactivating states with field
Having separate Activating and Deactivating classes forces clients to
handle four states instead of two, when many do not need to distinguish
beyond Disconnected and Connected. We can instead represent this
information with the "isTransitioning" and "status" fields. The "status"
field also allows us to be more descriptive about what is going on.

Plugins may still subclass Disconnected and Connected if it is useful.

Bug: 241591222
Test: existing
Change-Id: I24eb615df7c035167127782bac8f978ca2ec65ef