Read DirectAccess info from the catalog
This change includes a custom version of the testing api jar
with not-yet-visible functionality enabled.
It should be replaced with the public jar once one is published with the
requisite APIs enabled.
This lets us filter devices based on the flag in the catalog rather
than the hardcoded list we were using before.
Additionally, we now include the cloud project id in the catalog request.

Bug: http://go/da-table-row/8ZQoYr09W5BdKh7S1CaOYe
Test: pending (b/283121750)
Change-Id: I2d972cf9de5123b1aa6c3baaff6ad671218ceda4