Move tracking to invokeOnCompletion
device.awaitDisconnection() is called in the handle scope which is
cancelled when the user logs out. Any step after that point is not
executed. As a result we miss out on tracking this disconnect.
Additionally, with logout, stateFlow update was skipped leading to the
device panel showing error message in RDW.

Bug: N/A
Test: Added
Change-Id: I49a4ca5a163b202ba52d12e722ea47ea46c7198e