Track end reservation if reservation ended
When the reservation flow updates to a value resembling a terminated
reservation, multiple collectors are end their collection. One such
collector maps this reservation state to device provisioner reservation
state which updates the device's stateflow.
Device stateflow itself has a collector that tracks the reservation and
removes the device from the provisioner framework when the reservation
changes to ERROR or COMPLETED state. Removing the device results in the
device scope being cancelled.
The collector that tracks end reservation runs in device's scope. There
is a possibility that the collector did not finish before the scope was
cancelled resulting in a CancellationException in invokeOnCompletion.
This however needs to be tracked since the reservation ended
User logout will also cancel the device scope and result in the same
exception but does not have to be logged.
To overcome this, check if the reservation was closed before tracking it

Bug: 280523933
Test: Added
Change-Id: I75678bce64a7369344bddd082e4d404b1011e3f7