Verify template credentials more promptly when matching reservations
Whenever a GCP project is updated, a new list of reservations is
fetched to match templates with credentials. This is implemented by
checking the reservation's activationAction field. However, the
activationAction field may not be updated promptly, so it is possible
to match a reservation with a template incorrectly. This commit
filters the template list by matching it with the value
of activeDeviceInfoFlow directly to avoid potential race conditions.

Fixes: 293308110
Test: updated
Change-Id: I175df6f6c0f06718e3dfca2c364085bf5ab72c53