Add support to open a dialog to select templates to show in DM
Templates are cached after created in provisioner plugin.
They can be added to  or removed from the provisioner with the
dialog opened from cloud project selection panel.
Cached templates will be reused when added back.
The provisioner may hold a device handle without its template,
and the table row for the device handle will be removed with the
handle.

Bug: 290116162
Test: added in DirectAccessDeviceProvisionerTest
Change-Id: Ia937ef5389821736dae7118d1f278f547578e484