Track reservation end with correct state
When user force checks in the device, the reservation state changes to
FINISHED. This event was skipped in the project where the device was
force checked-in by checking the hasForceCheckedIn flag.
However, in a separate studio project where the device was connected,
this event gets counted as ERROR which should not be the case.
Track success metric when the reservation state is EXPIRED or FINISHED
and assign correct reason.

Bug: N/A
Test: Added
Change-Id: I4c5d8a9df72d6daf3fdb3b7f1480c070bcccf9fb