Fix device search in SelectDeviceDialog
Devices filtered out when searching were ignored if user hit "OK".
Since we saved the deviceRowDataList each time, we could not retrieve
the devices filtered out. To overcome this, don't change the list.
Instead, use it to apply the filters and return the list of selected
device when user hits "OK"

Fixes: 323814908
Fixes: 323804605
Fixes: 323798733
Test: Added
Change-Id: I9f65e9d3be185f30e9ffd2a25ef43870c1f49b95