Fix DirectAccessMultiProjectTest
The login coroutine was not guaranteed to run before the call to
selectProject(). This causes a race condition where the selectProject()
creates a cloud project manager only to be overridden nullified by the
login coroutine.
Mock DirectAccessPersistentStateComponent service and return a selected
project in order for the test to wait for the login coroutine to execute

Bug: 324482600
Test: enabled
Change-Id: Ie4fffc1cd4316f18631d7f521a0945cfda950313