Project System Token for test-recorder
Mostly implements helpers for the RecordingDialog, but also one or two
other helper functions relating to the project structure.

The Gradle implementation in this change is a straight copy of the
original, preserving as much as possible of the existing behaviour,
and (importantly) not attempting to address any of the issues present.

Bug: 147586218
Test: existing (no change in functionality)
Change-Id: Ic1c2d874804588d1e707028378806aa0cb49ca7a