Can fetch oem lab assets
To allow more flexibility and not just relying on
the built-in resources, we serve lab assets on gstatic.

The introduced OemLabsAssetsRegistry is to provide assets when asked,
it first tries to fetch from server. If it fails, we just fallback to
"unknown" assets. The download cache is set to expire in 1 day. The
local assets will be refreshed if it's modified since last fetch upon
asked

Bug: 368655162
Test: added
Change-Id: Ia4f0086dd38b3aed1ed6c105342215363166512d